Now the emulation of the cound is closer to real in C64. The modifications have to be extended the resto of emulators. The noise in C64 is not real yet, but it is improving!

+++ include/CORE/SoundVoice.hpp
@@ -204,9 +204,9 @@ namespace MCHEmul
 
 		/** All waves being part of the voice must "vibrate" at the same frequency. \n
 			Take care because no boundary analysis is done. */
-		unsigned short frequency () const
+		double frequency () const
 							{ return (_waves [0] -> frequency ()); /** could be any, as all have the same value. */ }
-		void setFrequency (unsigned short f)
+		void setFrequency (double f)
 							{ for (auto i : _waves) 
 								i -> setFrequency (f); }
